requirements.txt
Flask==3.0.0
hl7apy==1.3.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
